src/requirements.txt
accelerate==0.33.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
bitsandbytes==0.48.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.2.0
DateTime==5.5
debugpy==1.8.17
decorator==5.2.1
diffusers==0.31.0
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
fastapi==0.118.0
ffmpy==0.6.3
filelock==3.13.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.11.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.13
kiwisolver==1.4.9
langdetect==1.0.9
lightgbm==4.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==2.3.0
optuna==4.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pillow==11.0.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyngrok==7.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
ruff==0.14.0
safehttpx==0.1.6
safetensors==0.4.3
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.19.1
tomlkit==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
xgboost==3.0.5
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zope.interface==8.0.1
bitsandbytes==0.48.2
